# Eliminate exception allocation on parser hot path