--- a/ember-server/src/main/scala/org/http4s/ember/server/internal/ServerHelpers.scala
+++ b/ember-server/src/main/scala/org/http4s/ember/server/internal/ServerHelpers.scala
@@ -132,9 +132,11 @@ private[server] object ServerHelpers {
       durationToFinite(requestHeaderReceiveTimeout).fold(parse)(duration =>
         parse.timeoutTo(
           duration,
-          ApplicativeThrow[F].raiseError(
-            new java.util.concurrent.TimeoutException(
-              s"Timed Out on EmberServer Header Receive Timeout: $duration"))))
+          Concurrent[F].defer(
+            ApplicativeThrow[F].raiseError(new java.util.concurrent.TimeoutException(s"Timed Out on EmberServer Header Receive Timeout: $duration"))
+          )
+        )
+      )
 
     for {
       tmp <- parseWithHeaderTimeout