--- a/ember-server/src/main/scala/org/http4s/ember/server/internal/ServerHelpers.scala
+++ b/ember-server/src/main/scala/org/http4s/ember/server/internal/ServerHelpers.scala
@@ -133,10 +133,11 @@ private[server] object ServerHelpers {
         parse.timeoutTo(
           duration,
           Concurrent[F].defer(
-            ApplicativeThrow[F].raiseError(new java.util.concurrent.TimeoutException(s"Timed Out on EmberServer Header Receive Timeout: $duration"))
+            ApplicativeThrow[F].raiseError(
+              new java.util.concurrent.TimeoutException(
+                s"Timed Out on EmberServer Header Receive Timeout: $duration"))
           )
-        )
-      )
+        ))
 
     for {
       tmp <- parseWithHeaderTimeout