Strictness strikes again!

I ran a performance test against Ember Server with the change (after a warmup run)
```
$ hey -n 1000000 -c 50 http://127.0.0.1:8080/
```
Before:
```
Summary:
  Total:    58.2501 secs
  Slowest:    0.1386 secs
  Fastest:    0.0003 secs
  Average:    0.0029 secs
  Requests/sec:    17167.3672

  Total data:    14000000 bytes
  Size/request:    14 bytes

Response time histogram:
  0.000 [1]    |
  0.014 [995073]    |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.028 [3517]    |
  0.042 [964]    |
  0.056 [313]    |
  0.069 [101]    |
  0.083 [19]    |
  0.097 [7]    |
  0.111 [1]    |
  0.125 [3]    |
  0.139 [1]    |
```
After:
```
Summary:
  Total:    56.8599 secs
  Slowest:    0.1795 secs
  Fastest:    0.0002 secs
  Average:    0.0028 secs
  Requests/sec:    17587.0752

  Total data:    14000000 bytes
  Size/request:    14 bytes

Response time histogram:
  0.000 [1]    |
  0.018 [996375]    |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.036 [2582]    |
  0.054 [672]    |
  0.072 [235]    |
  0.090 [86]    |
  0.108 [29]    |
  0.126 [11]    |
  0.144 [4]    |
  0.162 [4]    |
  0.179 [1]    |
```
Noticed about a ~6% CPU drop as well